Refactor BIDS HED validation code into type-specific classes
